validacion de autorizacion para modificar y borrar tareas/actualizacion de mostrar tareas

+++ backend/controllers/productsControllers.js
@@ -2,9 +2,14 @@ const asyncHandler = require('express-async-handler')
 const Product = require('../models/productsModel')
 
 
-const getProducts = asyncHandler(async(req,res) =>{
+const getAllProducts = asyncHandler(async(req,res) =>{
     const products = await Product.find()
     res.status(200).json(products)
+})
+
+const getProducts = asyncHandler(async(req,res) =>{
+    const product = await Product.find({user: req.user.id})
+    res.status(200).json(product)
 })
 
 const createProducts = asyncHandler(async(req,res) =>{
@@ -33,8 +38,16 @@ const updateProducts = asyncHandler(async(req,res) =>{
         res.status(400)
         throw new Error ('El producto no existe')
     }
-    const productUpdate = await Product.findByIdAndUpdate(req.params.id , req.body,{new:true})
-    res.status(200).json(productUpdate)
+     if(product.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error("Usuario no autorizado");
+        
+     }else{
+        const productUpdate = await Product.findByIdAndUpdate(req.params.id , req.body,{new:true})
+        res.status(200).json(productUpdate)
+     }
+
+
 })
 
 const deleteProducts = asyncHandler(async(req,res) =>{
@@ -46,9 +59,16 @@ const deleteProducts = asyncHandler(async(req,res) =>{
         throw new Error ('El producto no existe')
     }
 
-    await product.deleteOne()
+    if(product.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error("Usuario no autorizado");
+        
+     }else{
+        await product.deleteOne()
 
-    res.status(200).json({id: req.params.id})
+         res.status(200).json({id: req.params.id})
+     }
+    
 })
 
 
@@ -56,5 +76,6 @@ module.exports ={
     getProducts,
     createProducts,
     updateProducts,
-    deleteProducts
+    deleteProducts,
+    getAllProducts
 }
